backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
Django==4.2.11
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.3.0
